Refactor management_console
